Move Integration fixtures to different file.
