+++ backend/code/iep/application/testing.py
@@ -13,13 +13,10 @@ from sqlalchemy.orm import sessionmaker
 
 from iep import app
 from iep.application.app import IAPConfigurator
+from iep.application.fixtures import FixturesMixin
 from iep.auth.drivers import command as user_command
 from iep.auth.drivers import query as user_query
 from iep.auth.models import User
-from iep.conventions.drivers import command as convention_command
-from iep.panels.drivers import command as panel_command
-from iep.grid.drivers import command as paneltime_command
-from iep.rooms.drivers import command as room_command
 
 
 class DeleteOnExit(object):
@@ -112,7 +109,7 @@ class IAPFixturesMixin(object):
         user_command.force_delete(uid)
 
 
-class IntegrationFixture(IAPFixturesMixin):
+class IntegrationFixture(IAPFixturesMixin, FixturesMixin):
     SESSION_CACHE = {}
     CONFIGURATOR_KEY = "app"
 
@@ -133,44 +130,6 @@ class IntegrationFixture(IAPFixturesMixin):
             self.SESSION_CACHE[self.CONFIGURATOR_KEY] = app
             self.after_configurator_start(app)
         return self.SESSION_CACHE[self.CONFIGURATOR_KEY]
-
-    @fixture
-    def convention1_uid(self):
-        uid = convention_command.save_new(name="convention 1")
-        yield uid
-        convention_command.force_delete(uid)
-
-    @fixture
-    def convention2_uid(self):
-        uid = convention_command.save_new(name="convention 2")
-        yield uid
-        convention_command.force_delete(uid)
-
-    @fixture
-    def panel1_uid(self, convention1_uid):
-        uid = panel_command.save_new(name="panel 1", convention_uid=convention1_uid)
-        yield uid
-        panel_command.force_delete(uid)
-
-    @fixture
-    def room1_uid(self, convention1_uid):
-        uid = room_command.save_new(name="room 1", convention_uid=convention1_uid)
-        yield uid
-        room_command.force_delete(uid)
-
-    @fixture
-    def room2_uid(self, convention1_uid):
-        uid = room_command.save_new(name="room 2", convention_uid=convention1_uid)
-        yield uid
-        room_command.force_delete(uid)
-
-    @fixture
-    def paneltime1_uid(self, convention1_uid, panel1_uid, room1_uid):
-        uid = paneltime_command.save_new(
-            convention_uid=convention1_uid, panel_uid=panel1_uid, room_uid=room1_uid
-        )
-        yield uid
-        paneltime_command.force_delete(uid)
 
 
 class WebTestFixture(IAPFixturesMixin, BaseWebTestFixture):
